crud services unit tests

+++ src/main/java/pl/pelikan/pelikanbe/hashtag/HashtagService.java
@@ -45,19 +45,21 @@ public class HashtagService {
 
     public void deleteHashtag(Long id) {
         Hashtag hashtag = getHashtagById(id);
-        if (hashtag.getHashtagCounters() != null) {
-            for (HashtagCounter counter : hashtag.getHashtagCounters()) {
-                counter.setHashtag(null);
+        if (hashtag != null) {
+            if (hashtag.getHashtagCounters() != null) {
+                for (HashtagCounter counter : hashtag.getHashtagCounters()) {
+                    counter.setHashtag(null);
+                }
             }
-        }
-        if (hashtag.getOffers() != null) {
-            for (Offer offer : hashtag.getOffers()) {
-                List<Hashtag> hashtags = offer.getHashtags();
-                hashtags.remove(hashtag);
-                offer.setHashtags(hashtags);
+            if (hashtag.getOffers() != null) {
+                for (Offer offer : hashtag.getOffers()) {
+                    List<Hashtag> hashtags = offer.getHashtags();
+                    hashtags.remove(hashtag);
+                    offer.setHashtags(hashtags);
+                }
             }
+            hashtagRepository.deleteById(id);
         }
-        hashtagRepository.deleteById(id);
     }
 
     public boolean existsHashtagById(Long id) {
